requirements.txt
markdown==3.8
typer==0.15.2
openai==1.76.0
langchain==0.3.24
langchain_openai==0.3.14
questionary==2.1.0
python-dotenv==1.1.0
maturin==1.8.3
requests==2.31.0
anthropic==0.38.0

